0.2.8.5 almost all sex actions converted

+++ src/com/lilithsthrone/game/character/body/Body.java
@@ -1165,7 +1165,7 @@ public class Body implements Serializable, XMLSaving {
 		importedPenis.testicle.internal = (Boolean.valueOf(testicles.getAttribute("internal")));
 		
 		try {
-			importedPenis.testicle.cumStored = Integer.valueOf(testicles.getAttribute("storedCum"));
+			importedPenis.testicle.cumStored = Float.valueOf(testicles.getAttribute("storedCum"));
 			importedPenis.testicle.cumRegeneration = Integer.valueOf(testicles.getAttribute("cumRegeneration"));
 			importedPenis.testicle.cumExpulsion = Integer.valueOf(testicles.getAttribute("cumExpulsion"));
 		} catch(Exception ex) {
